MST v4.11.1 Released

+++ Assets/MasterServerToolkit/MasterServer/Scripts/Modules/Rooms/MstRoomsClient.cs
@@ -99,7 +99,7 @@ namespace MasterServerToolkit.MasterServer
                 Password = password
             };
 
-            connection.SendMessage((ushort)MstOpCodes.GetRoomAccessRequest, roomAccessRequestPacket, (status, response) =>
+            connection.SendMessage(MstOpCodes.GetRoomAccessRequest, roomAccessRequestPacket, (status, response) =>
             {
                 if (status != ResponseStatus.Success)
                 {
